requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.2.0
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
pkg-resources==0.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
style==1.1.0
update==0.0.1
whitenoise==5.0.1
